Refactor to move new code out of _patches.
