+++ numba_dpex/dpnp_iface/_intrinsic.py
@@ -2,19 +2,242 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+from collections import namedtuple
+
 from llvmlite import ir as llvmir
+from llvmlite.ir import Constant
 from llvmlite.ir.types import DoubleType, FloatType
 from numba import types
+from numba.core import cgutils
+from numba.core import config as numba_config
 from numba.core.typing import signature
-from numba.extending import intrinsic
+from numba.extending import intrinsic, overload_classmethod
 from numba.np.arrayobj import (
-    _empty_nd_impl,
     _parse_empty_args,
     _parse_empty_like_args,
     get_itemsize,
+    make_array,
+    populate_array,
 )
 
 from numba_dpex.core.runtime import context as dpexrt
+from numba_dpex.core.types import DpnpNdArray
+from numba_dpex.core.types.dpctl_types import DpctlSyclQueue
+
+
+# XXX: The function should be moved into DpexTargetContext
+def make_queue(context, builder, arrtype):
+    """Utility function used for allocating a new queue.
+
+    This function will allocates a new queue (e.g. SYCL queue)
+    during LLVM code generation (lowering). Given a target context,
+    builder, array type, returns a LLVM value pointing at a numba-dpex
+    runtime allocated queue.
+
+    Args:
+        context (numba.core.base.BaseContext): Any of the context
+            derived from Numba's BaseContext
+            (e.g. `numba.core.cpu.CPUContext`).
+        builder (llvmlite.ir.builder.IRBuilder): The IR builder
+            from `llvmlite` for code generation.
+        arrtype (numba_dpex.core.types.dpnp_ndarray_type.DpnpNdArray):
+            Any of the array types derived from
+            `numba.core.types.nptypes.Array`,
+            e.g. `numba_dpex.core.types.dpnp_ndarray_type.DpnpNdArray`.
+            Refer to `numba_dpex.dpnp_iface._intrinsic.alloc_empty_arrayobj()`
+            function for details on how to construct this argument.
+
+    Returns:
+        ret (namedtuple): A namedtuple containing
+        `llvmlite.ir.instructions.ExtractValue` as `queue_ref`,
+        `llvmlite.ir.instructions.CastInstr` as `queue_address_ptr`
+        and `numba.core.pythonapi.PythonAPI` as `pyapi`.
+    """
+
+    pyapi = context.get_python_api(builder)
+    queue_struct_proxy = cgutils.create_struct_proxy(
+        DpctlSyclQueue(arrtype.queue)
+    )(context, builder)
+    queue_struct_ptr = queue_struct_proxy._getpointer()
+    queue_struct_voidptr = builder.bitcast(queue_struct_ptr, cgutils.voidptr_t)
+
+    address = context.get_constant(types.intp, id(arrtype.queue))
+    queue_address_ptr = builder.inttoptr(address, cgutils.voidptr_t)
+
+    dpexrtCtx = dpexrt.DpexRTContext(context)
+    dpexrtCtx.queuestruct_from_python(
+        pyapi, queue_address_ptr, queue_struct_voidptr
+    )
+
+    queue_struct = builder.load(queue_struct_ptr)
+    queue_ref = builder.extract_value(queue_struct, 1)
+
+    return_values = namedtuple(
+        "return_values", "queue_ref queue_address_ptr pyapi"
+    )
+    ret = return_values(queue_ref, queue_address_ptr, pyapi)
+
+    return ret
+
+
+def _empty_nd_impl(context, builder, arrtype, shapes):
+    """Utility function used for allocating a new array.
+
+    This function is used for allocating a new array during LLVM code
+    generation (lowering).  Given a target context, builder, array
+    type, and a tuple or list of lowered dimension sizes, returns a
+    LLVM value pointing at a Numba runtime allocated array.
+    """
+
+    arycls = make_array(arrtype)
+    ary = arycls(context, builder)
+
+    datatype = context.get_data_type(arrtype.dtype)
+    itemsize = context.get_constant(types.intp, get_itemsize(context, arrtype))
+
+    # compute array length
+    arrlen = context.get_constant(types.intp, 1)
+    overflow = Constant(llvmir.IntType(1), 0)
+    for s in shapes:
+        arrlen_mult = builder.smul_with_overflow(arrlen, s)
+        arrlen = builder.extract_value(arrlen_mult, 0)
+        overflow = builder.or_(overflow, builder.extract_value(arrlen_mult, 1))
+
+    if arrtype.ndim == 0:
+        strides = ()
+    elif arrtype.layout == "C":
+        strides = [itemsize]
+        for dimension_size in reversed(shapes[1:]):
+            strides.append(builder.mul(strides[-1], dimension_size))
+        strides = tuple(reversed(strides))
+    elif arrtype.layout == "F":
+        strides = [itemsize]
+        for dimension_size in shapes[:-1]:
+            strides.append(builder.mul(strides[-1], dimension_size))
+        strides = tuple(strides)
+    else:
+        raise NotImplementedError(
+            "Don't know how to allocate array with layout '{0}'.".format(
+                arrtype.layout
+            )
+        )
+
+    # Check overflow, numpy also does this after checking order
+    allocsize_mult = builder.smul_with_overflow(arrlen, itemsize)
+    allocsize = builder.extract_value(allocsize_mult, 0)
+    overflow = builder.or_(overflow, builder.extract_value(allocsize_mult, 1))
+
+    with builder.if_then(overflow, likely=False):
+        # Raise same error as numpy, see:
+        # https://github.com/numpy/numpy/blob/2a488fe76a0f732dc418d03b452caace161673da/numpy/core/src/multiarray/ctors.c#L1095-L1101    # noqa: E501
+        context.call_conv.return_user_exc(
+            builder,
+            ValueError,
+            (
+                "array is too big; `arr.size * arr.dtype.itemsize` is larger "
+                "than the maximum possible size.",
+            ),
+        )
+
+    (queue_ref, queue_ptr, pyapi) = make_queue(context, builder, arrtype)
+
+    # The queue_ref returned by make_queue if used to allocate a MemInfo
+    # object needs to be copied first. The reason for the copy is to
+    # properly manage the lifetime of the queue_ref object. The original
+    # object is owned by the parent dpctl.SyclQueue object and is deleted
+    # when the dpctl.SyclQueue is garbage collected. Whereas, the copied
+    # queue_ref is to be owned by the NRT_External_Allocator object of
+    # MemInfo, and its lifetime is tied to the MemInfo object.
+
+    dpexrtCtx = dpexrt.DpexRTContext(context)
+    queue_ref_copy = dpexrtCtx.copy_queue(builder, queue_ref)
+
+    usm_ty = arrtype.usm_type
+    usm_ty_map = {"device": 1, "shared": 2, "host": 3}
+    usm_type = context.get_constant(
+        types.uint64, usm_ty_map[usm_ty] if usm_ty in usm_ty_map else 0
+    )
+
+    args = (
+        context.get_dummy_value(),
+        allocsize,
+        usm_type,
+        queue_ref_copy,
+    )
+    mip = types.MemInfoPointer(types.voidptr)
+    arytypeclass = types.TypeRef(type(arrtype))
+    sig = signature(
+        mip,
+        arytypeclass,
+        types.intp,
+        types.uint64,
+        types.voidptr,
+    )
+    from numba_dpex.decorators import dpjit
+
+    op = dpjit(_call_usm_allocator)
+    fnop = context.typing_context.resolve_value_type(op)
+    # The _call_usm_allocator function will be compiled and added to
+    # registry when the get_call_type function is invoked.
+    fnop.get_call_type(context.typing_context, sig.args, {})
+    eqfn = context.get_function(fnop, sig)
+    meminfo = eqfn(builder, args)
+    pyapi.decref(queue_ptr)
+    data = context.nrt.meminfo_data(builder, meminfo)
+
+    intp_t = context.get_value_type(types.intp)
+    shape_array = cgutils.pack_array(builder, shapes, ty=intp_t)
+    strides_array = cgutils.pack_array(builder, strides, ty=intp_t)
+
+    populate_array(
+        ary,
+        data=builder.bitcast(data, datatype.as_pointer()),
+        shape=shape_array,
+        strides=strides_array,
+        itemsize=itemsize,
+        meminfo=meminfo,
+    )
+
+    return_values = namedtuple("return_values", "ary queue_ref")
+    ret = return_values(ary, queue_ref)
+
+    return ret
+
+
+@overload_classmethod(DpnpNdArray, "_usm_allocate")
+def _ol_array_allocate(cls, allocsize, usm_type, queue):
+    """Implements an allocator for dpnp.ndarrays."""
+
+    def impl(cls, allocsize, usm_type, queue):
+        return intrin_usm_alloc(allocsize, usm_type, queue)
+
+    return impl
+
+
+numba_config.DISABLE_PERFORMANCE_WARNINGS = 0
+
+
+def _call_usm_allocator(arrtype, size, usm_type, queue):
+    """Trampoline to call the intrinsic used for allocation"""
+    return arrtype._usm_allocate(size, usm_type, queue)
+
+
+numba_config.DISABLE_PERFORMANCE_WARNINGS = 1
+
+
+@intrinsic
+def intrin_usm_alloc(typingctx, allocsize, usm_type, queue):
+    """Intrinsic to call into the allocator for Array"""
+
+    def codegen(context, builder, signature, args):
+        [allocsize, usm_type, queue] = args
+        dpexrtCtx = dpexrt.DpexRTContext(context)
+        meminfo = dpexrtCtx.meminfo_alloc(builder, allocsize, usm_type, queue)
+        return meminfo
+
+    mip = types.MemInfoPointer(types.voidptr)  # return untyped pointer
+    sig = signature(mip, allocsize, usm_type, queue)
+    return sig, codegen
 
 
 def alloc_empty_arrayobj(context, builder, sig, args, is_like=False):
@@ -103,21 +326,6 @@ def fill_arrayobj(context, builder, ary, arrtype, queue_ref, fill_value):
         queue_ref,
     )
     return ary, arrtype
-
-
-@intrinsic
-def intrin_usm_alloc(typingctx, allocsize, usm_type, queue):
-    """Intrinsic to call into the allocator for Array"""
-
-    def codegen(context, builder, signature, args):
-        [allocsize, usm_type, queue] = args
-        dpexrtCtx = dpexrt.DpexRTContext(context)
-        meminfo = dpexrtCtx.meminfo_alloc(builder, allocsize, usm_type, queue)
-        return meminfo
-
-    mip = types.MemInfoPointer(types.voidptr)  # return untyped pointer
-    sig = signature(mip, allocsize, usm_type, queue)
-    return sig, codegen
 
 
 @intrinsic
